nitty picks
